Project rafactored, Logic separated in each files
