+++ inventory_system/app.py
@@ -1,184 +1,25 @@
-# app.py
-from flask import Flask, request, jsonify, session
-from flask_sqlalchemy import SQLAlchemy
-from werkzeug.security import generate_password_hash, check_password_hash
-from datetime import datetime, timedelta
-import os
+from flask import Flask
 
-app = Flask(__name__)
-app.secret_key = os.urandom(24)  # Secure secret key for sessions
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///inventory.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['PERMANENT_SESSION_LIFETIME'] = timedelta(minutes=30)  # Session expires after 30 minutes
+from auth import auth_bp
+from config import Config
+from extensions import db
+from inventory import inventory_bp
 
-db = SQLAlchemy(app)
 
-# Models
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
-    password = db.Column(db.String(120), nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    inventory_items = db.relationship('InventoryItem', backref='owner', lazy=True)
+def create_app():
+    app = Flask(__name__)
+    app.config.from_object(Config)
+    db.init_app(app)
 
-class InventoryItem(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100), nullable=False)
-    description = db.Column(db.String(500))
-    quantity = db.Column(db.Integer, nullable=False)
-    price = db.Column(db.Float, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+    # Register blueprints
+    app.register_blueprint(auth_bp)
+    app.register_blueprint(inventory_bp)
 
-# Authentication decorator
-def login_required(f):
-    def wrapper(*args, **kwargs):
-        if 'user_id' not in session:
-            return jsonify({'error': 'Authentication required'}), 401
-        return f(*args, **kwargs)
-    wrapper.__name__ = f.__name__
-    return wrapper
-
-# User Authentication Routes
-@app.route('/register', methods=['POST'])
-def register():
-    data = request.get_json()
-    
-    if not all(k in data for k in ['username', 'password', 'email']):
-        return jsonify({'error': 'Missing required fields'}), 400
-    
-    if User.query.filter_by(username=data['username']).first():
-        return jsonify({'error': 'Username already exists'}), 400
-    
-    if User.query.filter_by(email=data['email']).first():
-        return jsonify({'error': 'Email already exists'}), 400
-    
-    hashed_password = generate_password_hash(data['password'])
-    new_user = User(
-        username=data['username'],
-        password=hashed_password,
-        email=data['email']
-    )
-    
-    db.session.add(new_user)
-    db.session.commit()
-    
-    return jsonify({'message': 'User registered successfully'}), 201
-
-@app.route('/login', methods=['POST'])
-def login():
-    data = request.get_json()
-    
-    if not all(k in data for k in ['username', 'password']):
-        return jsonify({'error': 'Missing credentials'}), 400
-    
-    user = User.query.filter_by(username=data['username']).first()
-    
-    if user and check_password_hash(user.password, data['password']):
-        session.permanent = True
-        session['user_id'] = user.id
-        return jsonify({'message': 'Logged in successfully'}), 200
-    
-    return jsonify({'error': 'Invalid credentials'}), 401
-
-@app.route('/logout', methods=['POST'])
-@login_required
-def logout():
-    session.clear()
-    return jsonify({'message': 'Logged out successfully'}), 200
-
-# Inventory Routes
-@app.route('/inventory', methods=['POST'])
-@login_required
-def create_item():
-    data = request.get_json()
-    
-    if not all(k in data for k in ['name', 'quantity', 'price']):
-        return jsonify({'error': 'Missing required fields'}), 400
-    
-    new_item = InventoryItem(
-        name=data['name'],
-        description=data.get('description', ''),
-        quantity=data['quantity'],
-        price=data['price'],
-        user_id=session['user_id']
-    )
-    
-    db.session.add(new_item)
-    db.session.commit()
-    
-    return jsonify({
-        'message': 'Item created successfully',
-        'item_id': new_item.id
-    }), 201
-
-@app.route('/inventory', methods=['GET'])
-@login_required
-def get_items():
-    items = InventoryItem.query.filter_by(user_id=session['user_id']).all()
-    return jsonify([{
-        'id': item.id,
-        'name': item.name,
-        'description': item.description,
-        'quantity': item.quantity,
-        'price': item.price,
-        'created_at': item.created_at.isoformat()
-    } for item in items]), 200
-
-@app.route('/inventory/<int:item_id>', methods=['GET'])
-@login_required
-def get_item(item_id):
-    item = InventoryItem.query.filter_by(id=item_id, user_id=session['user_id']).first()
-    
-    if not item:
-        return jsonify({'error': 'Item not found'}), 404
-    
-    return jsonify({
-        'id': item.id,
-        'name': item.name,
-        'description': item.description,
-        'quantity': item.quantity,
-        'price': item.price,
-        'created_at': item.created_at.isoformat()
-    }), 200
-
-@app.route('/inventory/<int:item_id>', methods=['PUT'])
-@login_required
-def update_item(item_id):
-    item = InventoryItem.query.filter_by(id=item_id, user_id=session['user_id']).first()
-    
-    if not item:
-        return jsonify({'error': 'Item not found'}), 404
-    
-    data = request.get_json()
-    
-    if 'name' in data:
-        item.name = data['name']
-    if 'description' in data:
-        item.description = data['description']
-    if 'quantity' in data:
-        item.quantity = data['quantity']
-    if 'price' in data:
-        item.price = data['price']
-    
-    db.session.commit()
-    
-    return jsonify({'message': 'Item updated successfully'}), 200
-
-@app.route('/inventory/<int:item_id>', methods=['DELETE'])
-@login_required
-def delete_item(item_id):
-    item = InventoryItem.query.filter_by(id=item_id, user_id=session['user_id']).first()
-    
-    if not item:
-        return jsonify({'error': 'Item not found'}), 404
-    
-    db.session.delete(item)
-    db.session.commit()
-    
-    return jsonify({'message': 'Item deleted successfully'}), 200
-
-if __name__ == '__main__':
     with app.app_context():
         db.create_all()
+
+    return app
+
+if __name__ == '__main__':
+    app = create_app()
     app.run(debug=True)
