+++ inventory_system/auth.py
@@ -1,0 +1,66 @@
+from flask import Blueprint, jsonify, request, session
+from werkzeug.security import check_password_hash, generate_password_hash
+
+from extensions import db
+from models import User
+
+auth_bp = Blueprint('auth', __name__)
+
+# Login required decorator
+def login_required(f):
+    def wrapper(*args, **kwargs):
+        if 'user_id' not in session:
+            return jsonify({'error': 'Authentication required'}), 401
+        return f(*args, **kwargs)
+    wrapper.__name__ = f.__name__
+    return wrapper
+
+# Registration route
+@auth_bp.route('/register', methods=['POST'])
+def register():
+    data = request.get_json()
+    
+    if not all(k in data for k in ['username', 'password', 'email']):
+        return jsonify({'error': 'Missing required fields'}), 400
+    
+    if User.query.filter_by(username=data['username']).first():
+        return jsonify({'error': 'Username already exists'}), 400
+    
+    if User.query.filter_by(email=data['email']).first():
+        return jsonify({'error': 'Email already exists'}), 400
+    
+    hashed_password = generate_password_hash(data['password'])
+    new_user = User(
+        username=data['username'],
+        password=hashed_password,
+        email=data['email']
+    )
+    
+    db.session.add(new_user)
+    db.session.commit()
+    
+    return jsonify({'message': 'User registered successfully'}), 201
+
+# Login route
+@auth_bp.route('/login', methods=['POST'])
+def login():
+    data = request.get_json()
+    
+    if not all(k in data for k in ['username', 'password']):
+        return jsonify({'error': 'Missing credentials'}), 400
+    
+    user = User.query.filter_by(username=data['username']).first()
+    
+    if user and check_password_hash(user.password, data['password']):
+        session.permanent = True
+        session['user_id'] = user.id
+        return jsonify({'message': 'Logged in successfully'}), 200
+    
+    return jsonify({'error': 'Invalid credentials'}), 401
+
+# Logout route
+@auth_bp.route('/logout', methods=['POST'])
+@login_required
+def logout():
+    session.clear()
+    return jsonify({'message': 'Logged out successfully'}), 200
