+++ inventory_system/config.py
@@ -1,0 +1,9 @@
+import os
+from datetime import timedelta
+
+
+class Config:
+    SECRET_KEY = os.urandom(24)
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///inventory.db'
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    PERMANENT_SESSION_LIFETIME = timedelta(minutes=30)
